Integrates Locast Core; adds basic sync
This is the first pass at integrating Locast Core. This will replace the
use of SimpleContentProvider to provide the data layer (Locast Core has
a copy of it), so submodules are twiddled a bit here.

This adds basic sync - posting / downloading of content seems to roughly
work; adding accounts works; Android sync somewhat works.