Adds first pass at loading a rendered video for cards
As animated gifs aren't viable on Android, this instead uses a video
rendered by the server which it then downloads and loops.  As it turns
out, this is much more economical anyhow; the file size of the
downloaded video is about 1/10th the size of an equivalent animated GIF.

As one may imagine, using a lossy video format for video is much more
effective than a "lossless", 256-color hack.